+++ libs/opengl/src/CAssimpModel.cpp
@@ -28,9 +28,9 @@
 #endif
 #endif
 
+#include <mrpt/opengl/opengl_api.h>
 #include <mrpt/serialization/CArchive.h>
 #include <mrpt/system/filesystem.h>
-#include "opengl_internals.h"
 
 using namespace mrpt;
 using namespace mrpt::opengl;
