+++ libs/opengl/src/COpenGLVertexArrayObject.cpp
@@ -11,7 +11,7 @@
 
 #include <mrpt/core/exceptions.h>
 #include <mrpt/opengl/COpenGLVertexArrayObject.h>
-#include "opengl_internals.h"
+#include <mrpt/opengl/opengl_api.h>
 
 using namespace mrpt::opengl;
