+++ libs/opengl/src/CRenderizable.cpp
@@ -20,7 +20,7 @@
 
 #include <mutex>
 
-#include "opengl_internals.h"
+#include <mrpt/opengl/opengl_api.h>
 
 using namespace std;
 using namespace mrpt;
