+++ libs/opengl/src/CSetOfObjects.cpp
@@ -13,8 +13,8 @@
 #include <mrpt/opengl/CTexturedPlane.h>
 #include <mrpt/serialization/CArchive.h>
 
+#include <mrpt/opengl/opengl_api.h>
 #include <algorithm>
-#include "opengl_internals.h"
 
 using namespace mrpt;
 using namespace mrpt::opengl;
