+++ libs/opengl/src/CSetOfTexturedTriangles.cpp
@@ -12,7 +12,7 @@
 #include <mrpt/opengl/CSetOfTexturedTriangles.h>
 #include <mrpt/serialization/CArchive.h>
 
-#include "opengl_internals.h"
+#include <mrpt/opengl/opengl_api.h>
 
 using namespace std;
 using namespace mrpt;
