+++ libs/opengl/src/CText.cpp
@@ -12,8 +12,8 @@
 #include <mrpt/opengl/CText.h>
 #include <mrpt/serialization/CArchive.h>
 
+#include <mrpt/opengl/opengl_api.h>
 #include "gltext.h"
-#include "opengl_internals.h"
 
 using namespace mrpt;
 using namespace mrpt::opengl;
