+++ libs/opengl/src/CTextMessageCapable.cpp
@@ -11,7 +11,7 @@
 
 #include <mrpt/opengl/CTextMessageCapable.h>
 
-#include "opengl_internals.h"
+#include <mrpt/opengl/opengl_api.h>
 
 using namespace std;
 using namespace mrpt;
