+++ libs/opengl/src/CTexturedObject.cpp
@@ -10,10 +10,10 @@
 #include "opengl-precomp.h"  // Precompiled header
 
 #include <mrpt/opengl/CTexturedObject.h>
+#include <mrpt/opengl/opengl_api.h>
 #include <mrpt/serialization/CArchive.h>
 #include <mrpt/system/CTimeLogger.h>
 #include <memory>  // std::align
-#include "opengl_internals.h"
 
 using namespace mrpt;
 using namespace mrpt::opengl;
