+++ libs/tfest/src/se3_l2.cpp
@@ -57,7 +57,7 @@ using namespace std;
 //
 // 7. Scale computation (symmetric expression)
 //		s = sqrt( sum{i}( square(std::abs(pRi')) / sum{i}(
-//square(std::abs(pLi')) ) )
+// square(std::abs(pLi')) ) )
 //
 // 8. Translation computation (distance between the Right centroid and the
 // scaled and rotated Left centroid)
